feat(reactor-browser): add code splitting and move renown crypto out of reactor
- Add separate entry points for analytics, graphql, relational, and renown
  modules to enable tree-shaking of heavy deps (luxon, graphql-request,
  kysely, @renown/sdk) when not used by the consumer
- Add subpath exports for each split module in package.json
- Move initRenownCrypto/initConnectCrypto from reactor.ts into renown/crypto.ts
  so @renown/sdk is no longer a root-level import
- Move @electric-sql/pglite to devDependencies (type-only usage)
