+++ packages/reactor-browser/package.json
@@ -17,6 +17,22 @@
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
+    },
+    "./analytics": {
+      "types": "./dist/src/analytics/index.d.ts",
+      "import": "./dist/src/analytics/index.js"
+    },
+    "./graphql": {
+      "types": "./dist/src/graphql/client.d.ts",
+      "import": "./dist/src/graphql/client.js"
+    },
+    "./relational": {
+      "types": "./dist/src/relational/index.d.ts",
+      "import": "./dist/src/relational/index.js"
+    },
+    "./renown": {
+      "types": "./dist/src/renown/index.d.ts",
+      "import": "./dist/src/renown/index.js"
     }
   },
   "sideEffects": false,
@@ -29,7 +45,6 @@
     "codegen": "graphql-codegen --config codegen.ts && prettier --write src/graphql/gen/*.ts && eslint --fix src/graphql/gen/*.ts --no-warn-ignored"
   },
   "dependencies": {
-    "@electric-sql/pglite": "catalog:",
     "@powerhousedao/analytics-engine-browser": "workspace:*",
     "@powerhousedao/analytics-engine-core": "workspace:*",
     "@powerhousedao/reactor": "workspace:*",
@@ -49,6 +64,7 @@
     "slug": "^11.0.0"
   },
   "devDependencies": {
+    "@electric-sql/pglite": "catalog:",
     "@graphql-codegen/cli": "catalog:",
     "@graphql-codegen/typescript": "catalog:",
     "@graphql-codegen/typescript-graphql-request": "catalog:",
